License changes (this isn't software).
Brief, short, and to the point license. The gist is that anyone can
copy, use, and modify them so long as they give original credit to
me. Additionally, there is a disclosure that these notes are distributed
in good faith that they will not be used for dishonest purposes such
as cheating on an academic assignment.